Adding 'next' command to workshopper
Currenty, when a student finishes an exercise, the only way to print instructions for the next exercise is to fire up the workshopper menu, navigate down with down arrow key to the next exercise, and press enter.

The `next` command provides an easy way to quickly navigate to the next exercise:

```
workshopname next
```

This will print the instruction for the next exercise and updates workshopper's "current" exercise internally.